forge fmt

+++ src/ParadigmCTF2021/Secure/challenge/Setup.sol
@@ -8,7 +8,7 @@ interface WETH9 is ERC20Like {
 
 contract Setup {
     WETH9 public constant WETH = WETH9(0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2);
-    uint public constant WANT = 50 ether;
+    uint256 public constant WANT = 50 ether;
 
     Wallet public wallet;
 
@@ -21,9 +21,11 @@ contract Setup {
         wallet.allowModule(tokenModule);
 
         WETH.deposit{value: msg.value}();
-        WETH.approve(address(wallet), type(uint).max);
+        WETH.approve(address(wallet), type(uint256).max);
 
-        wallet.execModule(tokenModule, abi.encodeWithSelector(TokenModule.deposit.selector, WETH, address(this), msg.value));
+        wallet.execModule(
+            tokenModule, abi.encodeWithSelector(TokenModule.deposit.selector, WETH, address(this), msg.value)
+        );
     }
 
     function isSolved() public view returns (bool) {
